+++ lib/src/models/yust_doc.dart
@@ -95,14 +95,14 @@ abstract class YustDoc {
     String? userId,
     String? envId,
     DateTime? expiresAt,
-  })  : _id = id,
-        _createdAt = createdAt,
-        _createdBy = createdBy,
-        _modifiedAt = modifiedAt,
-        _modifiedBy = modifiedBy,
-        _userId = userId,
-        _envId = envId,
-        _expiresAt = expiresAt;
+  }) : _id = id,
+       _createdAt = createdAt,
+       _createdBy = createdBy,
+       _modifiedAt = modifiedAt,
+       _modifiedBy = modifiedBy,
+       _userId = userId,
+       _envId = envId,
+       _expiresAt = expiresAt;
 
   YustDoc.fromJson(Map<String, dynamic> json) : _id = '';
 
@@ -110,8 +110,13 @@ abstract class YustDoc {
 
   Map<String, dynamic> toExportJson() {
     final filteredJson = toJson();
-    Yust.helpers.removeKeysFromMap(filteredJson,
-        ['createdBy', 'modifiedBy', 'userId', 'envId', 'expiresAt']);
+    Yust.helpers.removeKeysFromMap(filteredJson, [
+      'createdBy',
+      'modifiedBy',
+      'userId',
+      'envId',
+      'expiresAt',
+    ]);
     return filteredJson;
   }