+++ lib/src/models/yust_file.g.dart
@@ -7,9 +7,9 @@ part of 'yust_file.dart';
 // **************************************************************************
 
 Map<String, dynamic> _$YustFileToJson(YustFile instance) => <String, dynamic>{
-      'name': instance.name,
-      'modifiedAt': instance.modifiedAt?.toIso8601String(),
-      'url': instance.url,
-      'hash': instance.hash,
-      'createdAt': instance.createdAt?.toIso8601String(),
-    };
+  'name': instance.name,
+  'modifiedAt': instance.modifiedAt?.toIso8601String(),
+  'url': instance.url,
+  'hash': instance.hash,
+  'createdAt': instance.createdAt?.toIso8601String(),
+};