+++ lib/src/services/yust_auth_service_shared.dart
@@ -49,8 +49,9 @@ class YustAuthServiceShared {
       ..domain = domain ?? email.split('@').last
       ..gender = gender
       ..lastLogin = DateTime.now()
-      ..lastLoginDomain =
-          Uri.base.scheme.contains('http') ? Uri.base.host : null;
+      ..lastLoginDomain = Uri.base.scheme.contains('http')
+          ? Uri.base.host
+          : null;
     await yust.dbService.saveDoc<YustUser>(Yust.userSetup, user);
     return user;
   }